backend/requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
fluent-logger==0.9.3
pytest==4.0.1
pytest-cov==2.5.1
pytest-django==3.4.4
pytest-pep8==1.0.6